1.0.5 - рефакторинг main, фикс url avatar (null), добавлен чекбокс загрузить еще фото в модалку
