+++ src/features/people/PersonAvatar.jsx
@@ -15,7 +15,18 @@ export default function PersonAvatar({
     setError(false);
 
     if (personId) {
+      // window.avatarAPI.getPath(personId).then((path) => {
+      //   const url = `${path}?t=${Date.now()}`;
+      //   console.log("🖼️ avatar path:", url);
+      //   if (isMounted) setSrc(url);
+      // });
       window.avatarAPI.getPath(personId).then((path) => {
+        if (!path) {
+          console.warn("⚠️ Аватар не найден для:", personId);
+          if (isMounted) setError(true);
+          return;
+        }
+
         const url = `${path}?t=${Date.now()}`;
         console.log("🖼️ avatar path:", url);
         if (isMounted) setSrc(url);
